streamlit/requirements.txt
numpy==1.24.4
Pillow
Requests==2.32.3
streamlit==1.40.1
streamlit_drawable_canvas==0.9.3
